Refactor CPTs to a single 'publicaciones' CPT
This commit refactors the custom post types `motor`, `regulador`, and `otro_repuesto` into a single CPT called `publicaciones`.

The changes include:
- A new `publicaciones` CPT that unifies the three old CPTs.
- A new `tipo` taxonomy to differentiate between "Motor", "Regulador", and "Otro Repuesto".
- The existing `categoria` and `marca` taxonomies are now associated with the new `publicaciones` CPT.
- The ACF field groups have been consolidated into a single group for `publicaciones`.
- The REST API routes have been updated to use `/publicaciones` instead of `/motors`.
- The Vue.js frontend has been updated to use the new API endpoints and data structure.
- A routing issue in the frontend that was causing an infinite loading state has been fixed.
- Import paths in the frontend have been fixed to point to the correct locations.

Note on frontend verification:
I was unable to build the Vue.js application due to issues with the `npm install` command in the provided environment. I have updated all the necessary files, but I could not verify that the frontend compiles and runs correctly. The user will need to run `npm install --legacy-peer-deps` and `npm run build` in the `wp-content/plugins/motorlan-api-vue/app` directory to build the frontend. I also had issues renaming directories, so the directory structure of the frontend is not yet updated.
