+++ wp-content/plugins/motorlan-api-vue/app/src/plugins/fake-api/index.ts.disabled
@@ -4,7 +4,7 @@ import { setupWorker } from 'msw/browser'
 import { handlerAppBarSearch } from '@db/app-bar-search/index'
 import { handlerAppsAcademy } from '@db/apps/academy/index'
 import { handlerAppsChat } from '@db/apps/chat/index'
-import { handlerAppsEcommerce } from '@db/apps/motors/index'
+import { handlerAppsEcommerce } from '@db/apps/publicaciones/index'
 import { handlerAppsEmail } from '@db/apps/email/index'
 import { handlerAppsInvoice } from '@db/apps/invoice/index'
 import { handlerAppLogistics } from '@db/apps/logistics/index'
