+++ wp-content/plugins/motorlan-api-vue/app/src/views/pages/cards/card-basic/CardBasic.vue
@@ -3,7 +3,7 @@ import avatar1 from '@images/avatars/avatar-1.png'
 import avatar2 from '@images/avatars/avatar-2.png'
 import avatar3 from '@images/avatars/avatar-3.png'
 import avatar4 from '@images/avatars/avatar-4.png'
-import eCommerce2 from '@images/motors/2.png'
+import eCommerce2 from '@images/publicaciones/2.png'
 import pages1 from '@images/pages/1.png'
 import pages2 from '@images/pages/2.png'
 import pages3 from '@images/pages/3.png'
